Added example Animation
This is an example animation. I have it exported as an FBX, but that's not necessary. The file contains the character mesh, it's skeleton, and the animations the character use. There is not currently a standard order the animations must be done in. The animation guide will include a list of what animations are required for a player character.